refactor(chat): unify the chat-box into one remote-capable component
ChatBox and RemoteChatBox were near-verbatim duplicates (same six props/builders)
split across the Core and Remote domains, with two equally-duplicated React
components. They're really one box whose remoteness is opt-in, so collapse them:

- One `Chat\Components\ChatBox` (`chat.box`) extending RemoteComponent — same
  origin by default, remote when source()/audience() are set. Delete the Core
  ChatBox and the Remote RemoteChatBox.
- One React `chat/components/chat-box.tsx` registered in the chat plugin; its
  transport and history seeding switch on `props.remote`. Delete the remote copy.
- Drop the dead `conversationId` prop (never sent or read). Conversation identity
  is carried by the rendered endpoints, matching how laravel/ai issues and
  resumes server-managed conversations.

Two mechanisms assumed "remote" meant a `remote.`-prefixed wire type; both are
generalized so a remote-capable `chat.box` works:
- RemoteSchemaResolver now treats `chat.box` as the external-URL chat type
  (fixing the prior silent defang of a chat.box arriving from a manifest).
- RemoteSourceTokenController drops the `remote.` prefix gate; the signed-ref
  verification is the real authorization.

`remote/` is left as generic fetch infrastructure (RemoteComponent + data-list).
