+++ resources/js/chat/transport.test.ts
@@ -126,7 +126,7 @@ describe("ndjsonChatTransport", () => {
       audience: "https://crm.example.test",
       source: "fixtures.crm",
       nodeId: "crm-chat",
-      nodeType: "remote.chat-box",
+      nodeType: "chat.box",
       ref: "sealed-ref",
       scopes: ["chat.write"],
       tokenEndpoint: "/custom/remote-tokens/fixtures.crm",
@@ -180,7 +180,7 @@ describe("ndjsonChatTransport", () => {
       audience: "https://crm.example.test",
       source: "fixtures.crm",
       nodeId: "crm-chat",
-      nodeType: "remote.chat-box",
+      nodeType: "chat.box",
       ref: "sealed-ref",
       scopes: ["chat.write"],
       tokenEndpoint: "/custom/remote-tokens/fixtures.crm",
