+++ workbench/app/Layouts/AppLayout.php
@@ -6,8 +6,8 @@ namespace Workbench\App\Layouts;
 use Illuminate\Http\Request;
 use Lattice\Lattice\Actions\Components\Action as ActionComponent;
 use Lattice\Lattice\Attributes\AsLayout;
+use Lattice\Lattice\Chat\Components\ChatBox;
 use Lattice\Lattice\Core\Components\Badge;
-use Lattice\Lattice\Core\Components\ChatBox;
 use Lattice\Lattice\Core\Components\FloatingPanel;
 use Lattice\Lattice\Core\Components\RawBlock;
 use Lattice\Lattice\Core\Components\Stack;
